requirements.txt

PyExecJS
requests